file rename and rename refs to EventFlatten
